séparation du code en plusieurs feuilles c
