+++ .gitignore
@@ -39,6 +39,6 @@ Thumbs.db
 #poubidou
 CodeC/Central_tech
 .idea
-.dat*
+
 .csv
 main
