+++ CodeC/macro.h
@@ -1,0 +1,14 @@
+//
+// Created by natha on 19/12/2024.
+//
+
+#ifndef MACRO_H
+#define MACRO_H
+
+#define NL printf("\n")
+#define MIN(a,b) (((a)<(b))?(a):(b))
+#define MAX(a,b) (((a)>(b))?(a):(b))
+
+enum type_station { DEFAUT, HVB, HVA, LV }; //defaut équivaut à nul, vide, etc
+
+#endif //MACRO_H
